# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: University Inventory Rental API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/rentals:
    get:
      summary: Get all rentals
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200': { description: OK }

    post:
      summary: Create a rental
      requestBody:
        required: true
      responses:
        '201': { description: Rental created }
  /api/rentals/{id}/return:
    post:
      summary: Return a rental item
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Returned }
        '400': { description: Already returned }
...